requirements.txt
absl-py==1.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.24
#dataclasses==0.8
deepdiff==5.7.0
docker-pycreds==0.4.0
docopt==0.6.2
enum34==1.1.10
estraces==1.7.1
fasteners==0.16.3
fastrand==1.3.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
glfw==2.1.0
gym==0.10.5
h5py==2.10.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.3
Jinja2==3.0.1
joblib==1.1.0
jsonpickle==1.5.2
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mock==4.0.3
mpyq==0.2.5
mujoco-py==2.0.2.8
#Mujoco200对应的Mujoco-py版本,下载mojuco200对应压缩包，解压，然后找教程。
#pip install mujoco-py==2.0.2.8
#使用python3.8
munch==2.5.0
numba==0.53.1
numpy==1.19.5
ordered-set==4.0.2
packaging==21.3
pathtools==0.1.2
Pillow==8.3.2
portpicker==1.5.0
promise==2.3
protobuf==3.17.3
psutil==5.9.0
py-cpuinfo==8.0.0
pycparser==2.20
pygame==2.1.2
pyglet==1.5.19
pyparsing==3.0.7
PySC2==3.0.0
python-dateutil==2.8.2
PyYAML==3.13
requests==2.26.0
s2clientprotocol==5.0.9.87702.0
s2protocol==5.0.9.87702.0
sacred==0.8.2
scared==0.8.13
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.11.1
setproctitle==1.2.3
shortuuid==1.0.11
six==1.16.0
sk-video==1.1.10
sklearn==0.0
smmap==5.0.0
tensorboard-logger==0.1.0
tensorboardX==2.4
threadpoolctl==3.1.0
torch==1.9.0
torchvision==0.10.0
trsfile==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.13
wandb==0.13.6
websocket-client==1.3.1
whichcraft==0.6.1
wrapt==1.14.0
zipp==3.6.0
